I converted evmone's benchmarks to StateTest format and they are available at https://github.com/ipsilon/evm-benchmarks. Here added as a git submodule. We are still discussing how exactly extract and maintain the benchmark suite outside of the evmone project.

Then I copied a lot of code from `TestState()` to create `BenchmarkState()` with the goal to use the lowest EVM execute function possible. I stopped at `EVM.Call()` but I believe even `Interpreter.Run()` is possible.

If some refactoring is done, some code can be shared between `TestState()` and `BenchmarkState()`. But the test runner is quite convoluted with `reflect` being used at some level...

```
go test -run - -bench='.*' ./tests/...
```