@chfast 2021-12-03
Comparison between `evm run --bench` and `go test -bench`.
```
name                             old time/op  new time/op  delta
micro/memory_grow_mload/nogrow    219µs ± 0%   230µs ± 1%   +4.81%  (p=0.000 n=18+18)
micro/memory_grow_mload/by1       230µs ± 1%   239µs ± 0%   +4.06%  (p=0.000 n=20+19)
micro/memory_grow_mload/by16      292µs ± 1%   283µs ± 0%   -3.11%  (p=0.000 n=19+20)
micro/memory_grow_mload/by32      367µs ± 1%   340µs ± 1%   -7.54%  (p=0.000 n=20+19)
micro/signextend/zero             390µs ± 1%   430µs ± 1%  +10.27%  (p=0.000 n=20+18)
micro/signextend/one              406µs ± 1%   445µs ± 1%   +9.46%  (p=0.000 n=20+20)
micro/jump_around/empty           145µs ± 1%   143µs ± 1%   -1.00%  (p=0.000 n=17+20)
micro/memory_grow_mstore/nogrow   445µs ± 0%   460µs ± 1%   +3.39%  (p=0.000 n=19+20)
micro/memory_grow_mstore/by1      460µs ± 1%   473µs ± 1%   +2.82%  (p=0.000 n=20+19)
micro/memory_grow_mstore/by16     539µs ± 0%   534µs ± 0%   -0.93%  (p=0.000 n=18+18)
micro/memory_grow_mstore/by32     607µs ± 1%   575µs ± 1%   -5.22%  (p=0.000 n=20+19)
micro/JUMPDEST_n0/empty          2.74ms ± 0%  3.13ms ± 1%  +14.07%  (p=0.000 n=19+20)
main/sha1_divs/empty              196µs ± 1%   202µs ± 0%   +3.07%  (p=0.000 n=20+20)
main/sha1_divs/1351              3.95ms ± 0%  4.12ms ± 0%   +4.16%  (p=0.000 n=20+19)
main/sha1_divs/2737              7.71ms ± 0%  8.05ms ± 0%   +4.42%  (p=0.000 n=19+20)
main/sha1_divs/5311              15.1ms ± 0%  15.7ms ± 0%   +4.31%  (p=0.000 n=18+19)
main/weierstrudel/0               471µs ± 0%   478µs ± 1%   +1.60%  (p=0.000 n=18+20)
main/weierstrudel/1               980µs ± 1%  1004µs ± 0%   +2.43%  (p=0.000 n=20+18)
main/weierstrudel/3              1.54ms ± 1%  1.58ms ± 0%   +2.73%  (p=0.000 n=20+20)
main/weierstrudel/9              3.18ms ± 0%  3.28ms ± 1%   +2.99%  (p=0.000 n=20+20)
main/weierstrudel/14             4.55ms ± 1%  4.71ms ± 1%   +3.38%  (p=0.000 n=18+20)
main/sha1_shifts/empty            151µs ± 0%   153µs ± 0%   +1.57%  (p=0.000 n=20+17)
main/sha1_shifts/1351            3.09ms ± 0%  3.19ms ± 0%   +3.24%  (p=0.000 n=14+19)
main/sha1_shifts/2737            6.02ms ± 0%  6.21ms ± 0%   +3.21%  (p=0.000 n=17+20)
main/sha1_shifts/5311            11.8ms ± 1%  12.2ms ± 1%   +3.19%  (p=0.000 n=19+20)
main/blake2b_huff/empty           100µs ± 0%   102µs ± 0%   +2.60%  (p=0.000 n=18+18)
main/blake2b_huff/2805nulls      1.90ms ± 1%  2.00ms ± 1%   +5.26%  (p=0.000 n=20+20)
main/blake2b_huff/5610nulls      3.70ms ± 0%  3.89ms ± 0%   +5.00%  (p=0.000 n=19+18)
main/blake2b_huff/8415nulls      5.43ms ± 1%  5.69ms ± 0%   +4.86%  (p=0.000 n=20+18)
main/blake2b_shifts/2805nulls    14.2ms ± 2%  15.3ms ± 1%   +7.48%  (p=0.000 n=20+20)
main/blake2b_shifts/5610nulls    28.1ms ± 0%  30.6ms ± 1%   +8.71%  (p=0.000 n=18+20)
main/blake2b_shifts/8415nulls    41.9ms ± 0%  45.6ms ± 0%   +9.00%  (p=0.000 n=16+18)
[Geo mean]                       1.40ms       2.42ms       +73.04%
```