@chfast 2021-12-08
I changed the implementation to use `Interpreter.Run()`. I did this to proof this is possible, but now I think this may be bad design because `StateTest` uses `EVM.Call()` and `evm run` uses `runtime.Call()` so we have 3 different entry points for EVM execution.

Using one of the `Call()` functions may be better. And we may improve the performance of this part later too.