@holiman 2021-12-10
> @holiman, please check if jumpdest analysis is cached for these runs.

Yeah, so apparently the `Run` does one analysis per benchmark, whereas the `Call` variants does `b.N`:
```
BenchmarkEVM/main/blake2b_huff/London/0/Run-6              10000            115034 ns/op                 1.000 analysis
BenchmarkEVM/main/blake2b_huff/London/1/Run-6                493           3029761 ns/op                 1.000 analysis
BenchmarkEVM/main/blake2b_huff/London/2/Run-6                236           5618322 ns/op                 1.000 analysis
BenchmarkEVM/main/blake2b_huff/London/3/Run-6                164           7499899 ns/op                 1.000 analysis
BenchmarkEVM/main/blake2b_huff/London/4/Run-6                 22          56706663 ns/op                 1.000 analysis
BenchmarkEVM/main/blake2b_huff/London/0/Call-6              8616            141017 ns/op              8616 analysis
BenchmarkEVM/main/blake2b_huff/London/1/Call-6               469           2526879 ns/op               469.0 analysis
BenchmarkEVM/main/blake2b_huff/London/2/Call-6               238           4763046 ns/op               238.0 analysis
BenchmarkEVM/main/blake2b_huff/London/3/Call-6               160           7526510 ns/op               160.0 analysis
BenchmarkEVM/main/blake2b_huff/London/4/Call-6                19          54531691 ns/op                19.00 analysis
```