@holiman 2021-12-10
It might be possible to fix this via a hack. Instead lf 
```
			sender := vm.AccountRef(msg.From())
			b.ResetTimer()

			for n := 0; n < b.N; n++ {
				// Execute the message.
				snapshot := statedb.Snapshot()
				_, _, err = evm.Call(sender, *msg.To(), msg.Data(), msg.Gas(), msg.Value())

```
You make the `sender` a `Contract`. That way, the `c.jumpdests` will be populated (and reused) via the child call.