@chfast 2021-12-15
> I suppose `Call` might be 'correcter', but I also think we should avoid including jumpdest analysis time in these benchmarks

Maybe the way is to be have API with more explicit code analysis handling.