@holiman 2021-12-21
> Maybe the way is to be have API with more explicit code analysis handling.

Do you have any proposal? I suggested a way to hack around it -- do you think there's some better way to do it? I don't really have any good alternatives.