@chfast 2021-12-21
Wrapping with another `Call` seems not easy and wrong because this explicitly modifies the StateTest structure. But I can try to do this to see how this work.

Alternative is to have more explicit code analysis cache management, i.e. analyze the code with one API outside of the loop and pass it to `Call()` inside the loop. So far the `Contract` abstraction which looks up it's sender seems very convoluted. But this seems to be longer work and we need EVM benchmark now.