@chfast 2021-12-21
BTW, how do you collect "analysis" stats for benchmarks?