@chfast 2021-12-21
I did one more commit which looks a bit better (and looks better IMHO) but in general the loop seems to be getting little faster but this is not visible in benchmarks doing more than just JUMPDEST.

```
name                                   old time/op    new time/op    delta
SimpleLoop/staticcall-identity-100M-8     203ms ± 0%     205ms ± 2%    ~     (p=0.156 n=9+10)
SimpleLoop/call-identity-100M-8           283ms ± 1%     282ms ± 1%  -0.47%  (p=0.029 n=10+10)
SimpleLoop/loop-100M-8                    281ms ± 1%     276ms ± 1%  -1.79%  (p=0.000 n=10+10)
SimpleLoop/call-nonexist-100M-8           631ms ± 2%     624ms ± 0%  -1.06%  (p=0.000 n=10+9)
SimpleLoop/call-EOA-100M-8                279ms ± 1%     274ms ± 1%  -1.62%  (p=0.000 n=9+10)
SimpleLoop/call-reverting-100M-8          557ms ± 0%     559ms ± 1%    ~     (p=0.075 n=10+10)
[Geo mean]                                341ms          339ms       -0.65%
```

```
name                                  old time/op  new time/op  delta                                                                                                                                                             
micro/memory_grow_mload/nogrow         212µs ± 1%   208µs ± 1%   -2.04%  (p=0.000 n=19+20)                                                                                                                                        
micro/memory_grow_mload/by1            223µs ± 1%   218µs ± 1%   -1.86%  (p=0.000 n=18+20)                                                                                                                                        
micro/memory_grow_mload/by16           285µs ± 0%   281µs ± 1%   -1.45%  (p=0.000 n=18+20)                                                                                                                                        
micro/memory_grow_mload/by32           358µs ± 1%   354µs ± 1%   -1.24%  (p=0.000 n=20+20)                                                                                                                                        
micro/signextend/zero                  355µs ± 0%   307µs ± 1%  -13.27%  (p=0.000 n=20+17)                                                                                                                                        
micro/signextend/one                   369µs ± 1%   320µs ± 1%  -13.35%  (p=0.000 n=19+20)                                                                                                                                        
micro/loop_with_many_jumpdests/empty  1.69µs ± 2%  1.68µs ± 2%     ~     (p=0.433 n=20+20)                                                                                                                                        
micro/jump_around/empty                135µs ± 1%   132µs ± 1%   -1.80%  (p=0.000 n=20+20)                                                                                                                                        
micro/memory_grow_mstore/nogrow        449µs ± 0%   438µs ± 1%   -2.49%  (p=0.000 n=18+20)                                                                                                                                        
micro/memory_grow_mstore/by1           464µs ± 1%   452µs ± 1%   -2.62%  (p=0.000 n=20+20)                                                                                                                                        
micro/memory_grow_mstore/by16          523µs ± 0%   535µs ± 0%   +2.22%  (p=0.000 n=20+20)                                                                                                                                        
micro/memory_grow_mstore/by32          602µs ± 1%   600µs ± 1%   -0.29%  (p=0.015 n=20+19)                                                                                                                                        
micro/JUMPDEST_n0/empty               2.58ms ± 0%  2.34ms ± 0%   -9.39%  (p=0.000 n=18+17)                                                                                                                                        
main/sha1_divs/empty                   176µs ± 0%   179µs ± 1%   +1.26%  (p=0.000 n=19+20)                                                                                                                                        
main/sha1_divs/1351                   3.55ms ± 1%  3.59ms ± 1%   +1.10%  (p=0.000 n=19+19)
main/sha1_divs/2737                   6.92ms ± 0%  7.00ms ± 0%   +1.17%  (p=0.000 n=17+18)
main/sha1_divs/5311                   13.5ms ± 0%  13.7ms ± 1%   +1.33%  (p=0.000 n=20+20)
main/weierstrudel/0                    446µs ± 0%   444µs ± 0%   -0.62%  (p=0.000 n=20+19)
main/weierstrudel/1                    923µs ± 0%   913µs ± 1%   -1.16%  (p=0.000 n=18+19)
main/weierstrudel/3                   1.45ms ± 0%  1.44ms ± 0%   -0.89%  (p=0.000 n=19+17)
main/weierstrudel/9                   3.03ms ± 1%  2.99ms ± 1%   -1.25%  (p=0.000 n=20+19)
main/weierstrudel/14                  4.33ms ± 0%  4.28ms ± 1%   -1.04%  (p=0.000 n=19+20)
main/sha1_shifts/empty                 132µs ± 1%   137µs ± 0%   +3.54%  (p=0.000 n=20+20)
main/sha1_shifts/1351                 2.71ms ± 1%  2.79ms ± 0%   +2.96%  (p=0.000 n=20+20)
main/sha1_shifts/2737                 5.28ms ± 0%  5.43ms ± 0%   +2.91%  (p=0.000 n=18+18)
main/sha1_shifts/5311                 10.3ms ± 0%  10.6ms ± 0%   +3.10%  (p=0.000 n=18+19)
main/blake2b_huff/empty               95.7µs ± 0%  96.3µs ± 0%   +0.67%  (p=0.000 n=17+18)
main/blake2b_huff/2805nulls           1.85ms ± 0%  1.86ms ± 0%   +0.33%  (p=0.000 n=19+20)
main/blake2b_huff/5610nulls           3.59ms ± 0%  3.61ms ± 0%   +0.40%  (p=0.000 n=17+19)
main/blake2b_huff/8415nulls           5.26ms ± 0%  5.28ms ± 1%   +0.40%  (p=0.000 n=19+19)
main/blake2b_shifts/2805nulls         13.1ms ± 2%  13.2ms ± 2%   +1.13%  (p=0.001 n=19+17)
main/blake2b_shifts/5610nulls         25.9ms ± 2%  26.2ms ± 1%   +1.53%  (p=0.000 n=17+19)
main/blake2b_shifts/8415nulls         38.4ms ± 1%  39.2ms ± 2%   +2.01%  (p=0.000 n=17+18)
[Geo mean]                            1.07ms       1.06ms        -0.96%
```