@chfast 2022-01-03
I compared two implementation variants: Run (old) and Call (new). The difference is small and does not which variant is faster. So I'm for staying with "Call" variant.

```
name                                               old time/op  new time/op  delta
EVM/main/blake2b_huff/London/0/Run-8               85.1µs ± 1%  85.8µs ± 1%  +0.84%  (p=0.000 n=9+9)
EVM/main/blake2b_huff/London/1/Run-8               1.84ms ± 1%  1.85ms ± 0%  +0.21%  (p=0.035 n=9+10)
EVM/main/blake2b_huff/London/2/Run-8               3.59ms ± 0%  3.60ms ± 0%  +0.31%  (p=0.003 n=8+8)
EVM/main/blake2b_huff/London/3/Run-8               5.38ms ± 0%  5.28ms ± 0%  -1.82%  (p=0.000 n=8+10)
EVM/main/blake2b_huff/London/4/Run-8               41.4ms ± 2%  41.0ms ± 0%    ~     (p=0.481 n=10+10)
EVM/main/blake2b_shifts/London/0/Run-8             12.7ms ± 0%  12.7ms ± 0%    ~     (p=0.274 n=8+10)
EVM/main/blake2b_shifts/London/1/Run-8             25.4ms ± 1%  25.6ms ± 1%    ~     (p=0.052 n=10+10)
EVM/main/blake2b_shifts/London/2/Run-8             37.8ms ± 0%  37.9ms ± 0%  +0.35%  (p=0.001 n=9+10)
EVM/main/blake2b_shifts/London/3/Run-8              294ms ± 1%   294ms ± 1%    ~     (p=0.190 n=10+10)
EVM/main/sha1_divs/London/0/Run-8                   175µs ± 2%   173µs ± 1%    ~     (p=0.105 n=10+10)
EVM/main/sha1_divs/London/1/Run-8                  3.59ms ± 2%  3.54ms ± 1%  -1.43%  (p=0.023 n=10+10)
EVM/main/sha1_divs/London/2/Run-8                  7.03ms ± 2%  6.91ms ± 1%  -1.64%  (p=0.015 n=10+10)
EVM/main/sha1_divs/London/3/Run-8                  13.9ms ± 0%  13.5ms ± 1%  -2.81%  (p=0.000 n=9+10)
EVM/main/sha1_divs/London/4/Run-8                   166ms ± 3%   165ms ± 1%    ~     (p=0.315 n=10+10)
EVM/main/sha1_shifts/London/0/Run-8                 130µs ± 3%   128µs ± 1%    ~     (p=0.165 n=10+10)
EVM/main/sha1_shifts/London/1/Run-8                2.69ms ± 9%  2.66ms ± 1%    ~     (p=0.143 n=10+10)
EVM/main/sha1_shifts/London/2/Run-8                5.24ms ± 3%  5.18ms ± 0%    ~     (p=0.481 n=10+10)
EVM/main/sha1_shifts/London/3/Run-8                10.4ms ± 4%  10.2ms ± 1%    ~     (p=0.143 n=10+10)
EVM/main/sha1_shifts/London/4/Run-8                 125ms ± 3%   123ms ± 0%    ~     (p=1.000 n=10+10)
EVM/main/weierstrudel/London/0/Run-8                427µs ± 0%   428µs ± 0%  +0.37%  (p=0.001 n=10+10)
EVM/main/weierstrudel/London/1/Run-8                891µs ± 1%   891µs ± 1%    ~     (p=0.780 n=10+9)
EVM/main/weierstrudel/London/2/Run-8               1.41ms ± 0%  1.41ms ± 0%    ~     (p=0.971 n=10+10)
EVM/main/weierstrudel/London/3/Run-8               2.97ms ± 0%  2.96ms ± 1%  -0.33%  (p=0.019 n=10+10)
EVM/main/weierstrudel/London/4/Run-8               4.25ms ± 1%  4.24ms ± 0%    ~     (p=0.122 n=10+8)
EVM/micro/JUMPDEST_n0/London/0/Run-8               2.58ms ± 0%  2.58ms ± 0%  +0.19%  (p=0.002 n=9+9)
EVM/micro/jump_around/London/0/Run-8                116µs ± 1%   117µs ± 1%  +0.56%  (p=0.003 n=10+10)
EVM/micro/loop_with_many_jumpdests/London/0/Run-8  39.1ms ± 1%  39.1ms ± 0%    ~     (p=0.447 n=10+9)
EVM/micro/memory_grow_mload/London/0/Run-8          211µs ± 2%   209µs ± 0%    ~     (p=0.447 n=10+9)
EVM/micro/memory_grow_mload/London/1/Run-8          218µs ± 2%   221µs ± 1%  +1.36%  (p=0.001 n=9+10)
EVM/micro/memory_grow_mload/London/2/Run-8          262µs ± 1%   264µs ± 0%  +0.87%  (p=0.000 n=10+10)
EVM/micro/memory_grow_mload/London/3/Run-8          318µs ± 2%   321µs ± 1%    ~     (p=0.075 n=10+10)
EVM/micro/memory_grow_mstore/London/0/Run-8         434µs ± 1%   436µs ± 0%  +0.32%  (p=0.040 n=9+9)
EVM/micro/memory_grow_mstore/London/1/Run-8         448µs ± 1%   451µs ± 0%  +0.62%  (p=0.001 n=9+10)
EVM/micro/memory_grow_mstore/London/2/Run-8         494µs ± 0%   497µs ± 0%  +0.55%  (p=0.000 n=10+10)
EVM/micro/memory_grow_mstore/London/3/Run-8         547µs ± 1%   554µs ± 1%  +1.24%  (p=0.000 n=10+10)
EVM/micro/signextend/London/0/Run-8                 318µs ± 0%   319µs ± 0%  +0.38%  (p=0.008 n=10+9)
EVM/micro/signextend/London/1/Run-8                 332µs ± 0%   334µs ± 0%  +0.38%  (p=0.000 n=10+9)
[Geo mean]                                         2.28ms       2.28ms       -0.27%
```